refactor: replace DeployContractSection with InvestorContractStep and update related documentation
