+++ app/src/components/sections/TeamView/forms/__tests__/ContinueAddTeamForm.spec.ts
@@ -4,7 +4,7 @@ import { ref } from 'vue'
 import ContinueAddTeamForm from '@/components/sections/TeamView/forms/ContinueAddTeamForm.vue'
 import type { Team } from '@/types/team'
 import { createTestingPinia } from '@pinia/testing'
-import DeployContractSection from '@/components/sections/TeamView/forms/DeployContractSection.vue'
+import InvestorContractStep from '@/components/sections/TeamView/forms/InvestorContractStep.vue'
 
 // Hoisted mocks for Safe composables
 const { mockUseSafe } = vi.hoisted(() => ({
@@ -145,7 +145,7 @@ describe.skip('ContinueAddTeamForm', () => {
     // Expect the error message to not be displayed
     expect(wrapper.find("[data-test='share-name-error']").exists()).toBeFalsy()
 
-    const deployComponent = wrapper.findComponent(DeployContractSection)
+    const deployComponent = wrapper.findComponent(InvestorContractStep)
     expect(wrapper.emitted('done')).toBeFalsy()
 
     await deployComponent.vm.$emit('contractDeployed')
@@ -214,7 +214,7 @@ describe.skip('ContinueAddTeamForm', () => {
       }
     })
 
-    const deployComponent = wrapper.findComponent(DeployContractSection)
+    const deployComponent = wrapper.findComponent(InvestorContractStep)
 
     expect(wrapper.emitted('done')).toBeFalsy()
